cleanup: remove unused code and stale tests
