+++ manager/pkg/service/sandbox_service_claim_test.go
@@ -265,76 +265,6 @@ func TestClaimIdlePodRequestsDeleteAfterNetworkApplyFailure(t *testing.T) {
 	}
 }
 
-func TestWaitForPodReadyWaitsUntilReady(t *testing.T) {
-	pod := newClaimTestPod("ns-a", "cold-pod", "template-a", false)
-	indexer := newClaimTestPodIndexer(t, pod)
-	svc := &SandboxService{
-		podLister: corelisters.NewPodLister(indexer),
-		config: SandboxServiceConfig{
-			ProcdInitTimeout: 100 * time.Millisecond,
-		},
-	}
-
-	go func() {
-		time.Sleep(80 * time.Millisecond)
-		updated := pod.DeepCopy()
-		updated.Status.Conditions[0].Status = corev1.ConditionTrue
-		updated.Status.Conditions[0].LastTransitionTime = metav1.NewTime(time.Now().UTC())
-		if err := indexer.Update(updated); err != nil {
-			t.Errorf("update pod: %v", err)
-		}
-	}()
-
-	readyPod, err := svc.waitForPodReady(context.Background(), pod.Namespace, pod.Name)
-	if err != nil {
-		t.Fatalf("waitForPodReady() error = %v", err)
-	}
-	if !controller.IsPodReady(readyPod) {
-		t.Fatalf("waitForPodReady() returned pod that is not ready")
-	}
-}
-
-func TestWaitForPodReadyTimesOut(t *testing.T) {
-	pod := newClaimTestPod("ns-a", "cold-pod", "template-a", false)
-	svc := &SandboxService{
-		podLister: newClaimTestPodLister(t, pod),
-	}
-
-	ctx, cancel := context.WithCancel(context.Background())
-	cancel()
-
-	_, err := svc.waitForPodReady(ctx, pod.Namespace, pod.Name)
-	if err == nil {
-		t.Fatal("waitForPodReady() error = nil, want timeout")
-	}
-}
-
-func TestWaitForPodReadyWaitsForPodToAppear(t *testing.T) {
-	indexer := cache.NewIndexer(cache.MetaNamespaceKeyFunc, cache.Indexers{cache.NamespaceIndex: cache.MetaNamespaceIndexFunc})
-	svc := &SandboxService{
-		podLister: corelisters.NewPodLister(indexer),
-		config: SandboxServiceConfig{
-			ProcdInitTimeout: 100 * time.Millisecond,
-		},
-	}
-
-	pod := newClaimTestPod("ns-a", "cold-pod", "template-a", true)
-	go func() {
-		time.Sleep(80 * time.Millisecond)
-		if err := indexer.Add(pod); err != nil {
-			t.Errorf("add pod: %v", err)
-		}
-	}()
-
-	readyPod, err := svc.waitForPodReady(context.Background(), pod.Namespace, pod.Name)
-	if err != nil {
-		t.Fatalf("waitForPodReady() error = %v", err)
-	}
-	if readyPod.Name != pod.Name {
-		t.Fatalf("waitForPodReady() returned %q, want %q", readyPod.Name, pod.Name)
-	}
-}
-
 func TestWaitForPodClaimReadyUsesSandboxReadinessWithoutPodReady(t *testing.T) {
 	pod := newClaimReadyTestPod("ns-a", "cold-pod", "template-a")
 	indexer := newClaimTestPodIndexer(t, pod)
